+++ test/get-items.ts
@@ -2,6 +2,7 @@ import type { StorageKey, StorageProvider } from "@storage/types";
 import { canGetItems, getItems } from "@storage/fns/get-items";
 import { assertArrayIncludes } from "@std/assert/array-includes";
 import { TEST_ITEMS } from "./fixtures.ts";
+import { assertEquals } from "@std/assert/equals";
 
 /**
  * Test the {@linkcode getItems} function of the given storage module,
@@ -17,10 +18,23 @@ export async function testGetItems(
     name: "getItems",
     fn: async () => {
       const keys = Array.from(items).map(([key, _value]) => key);
-      assertArrayIncludes(
-        await Array.fromAsync(getItems(store, keys)),
-        Array.from(items),
-      );
+      const foundItems = await Array.fromAsync(getItems(store, keys));
+      const expectedItems = Array.from(items);
+      assertArrayIncludes(foundItems, expectedItems);
+      assertEquals(foundItems.length, expectedItems.length);
+    },
+  });
+
+  await t.step({
+    ignore: !canGetItems(store),
+    name: "getItems - keys are deduplicated",
+    fn: async () => {
+      const keys = Array.from(items).map(([key, _value]) => key);
+      const duplicateKeys = [...keys, ...keys];
+      const foundItems = await Array.fromAsync(getItems(store, duplicateKeys));
+      const expectedItems = Array.from(items);
+      assertArrayIncludes(foundItems, expectedItems);
+      assertEquals(foundItems.length, expectedItems.length);
     },
   });
 }